refactor(core): derive the pricing projection from its shape and retire dead helpers
Cleanup pass over the previous commit.

`projectModelCallPricingRecord` named all 21 fields a third time, next to the
interface and the shape. Only the shape is exhaustively checked at compile
time, so an added optional field would have been silently dropped. It now
projects through `pickShape`, a new `record-schema` helper that keeps the keys
a shape allows — the shape becomes the one list, and the special-cased token
loop goes with it.

The migration selected every row into JS and used a thrown decode error to mean
"already narrow", so its steady state was a full-table round trip that could
never do anything. `schemaVersion` discriminates the two shapes in both
directions — required on an attempt, rejected by the pricing decoder — so
SQLite now selects exactly the rows still to fold, and the remaining catch
means what it says.

`assertPricingInvariants` took a label only to vary an error prefix no caller
reads. `sumModelCallCostUsd` and `isModelCallAttempt` had no production
consumers; the former was being retyped here for nobody.

Records the invariant this all rests on where it can be found: deleting a
Session cascades its runs and events but deliberately leaves the ledger rows,
so for a deleted Session the projection is the last copy of that spend.

Generated-by: Claude Code
