+++ packages/core/src/__tests__/model-call-attempt.test.ts
@@ -29,7 +29,6 @@ import {
   groupModelCallAttempts,
   projectModelCallPricingRecord,
   settledAttempt,
-  sumModelCallCostUsd,
   summarizeModelCallCoverage,
   type ModelCallAttempt,
 } from '../model-call-attempt.js';
@@ -361,35 +360,23 @@ describe('ModelCallAttempt projections', () => {
     });
   });
 
-  test('cost sum reports the qualifying coverage alongside the total', () => {
-    const { costUsd, coverage } = sumModelCallCostUsd([
-      attempt({ attemptId: 'a', costUsd: 0.004 }),
-      attempt({ attemptId: 'b', costUsd: 0.006 }),
-      attempt({ attemptId: 'c', costBasis: 'unpriced', costUsd: undefined }),
-    ]);
-    assert.equal(Math.round(costUsd * 1000) / 1000, 0.01);
-    assert.equal(coverage.unpricedAttempts, 1);
-  });
-
-  test('a replayed attemptId is counted once through sum and coverage', () => {
+  test('a replayed attemptId is counted once through coverage and grouping', () => {
     const stream = [
       attempt({ attemptId: 'a', logicalCallId: 'call-1', costUsd: 0.004 }),
       attempt({ attemptId: 'b', logicalCallId: 'call-2', costUsd: 0.006 }),
       attempt({ attemptId: 'a', logicalCallId: 'call-1', costUsd: 0.005 }),
     ];
-    const { costUsd, coverage } = sumModelCallCostUsd(stream);
-    assert.equal(Math.round(costUsd * 1000) / 1000, 0.011);
+    const coverage = summarizeModelCallCoverage(stream);
     assert.equal(coverage.attempts, 2);
     assert.equal(coverage.pricedAttempts, 2);
-    assert.equal(summarizeModelCallCoverage(stream).attempts, 2);
     assert.equal(groupModelCallAttempts(stream).length, 2);
   });
 
   test('a genuinely free priced call is distinguishable from an unpriced one', () => {
-    const free = attempt({ attemptId: 'free', costBasis: 'priced', costUsd: 0 });
-    const unknown = attempt({ attemptId: 'unknown', costBasis: 'unpriced', costUsd: undefined });
-    const { costUsd, coverage } = sumModelCallCostUsd([free, unknown]);
-    assert.equal(costUsd, 0);
+    const coverage = summarizeModelCallCoverage([
+      attempt({ attemptId: 'free', costBasis: 'priced', costUsd: 0 }),
+      attempt({ attemptId: 'unknown', costBasis: 'unpriced', costUsd: undefined }),
+    ]);
     assert.equal(coverage.pricedAttempts, 1);
     assert.equal(coverage.unpricedAttempts, 1);
   });
@@ -453,6 +440,8 @@ describe('the pricing record a Usage read model stores', () => {
   });
 
   test('holds a stored row to the same cost invariants as the authority', () => {
+    // The rules themselves are covered against the authority codec; this only
+    // proves the read model is wired to the same ones.
     assert.throws(
       () =>
         decodeModelCallPricingRecord({
@@ -460,22 +449,6 @@ describe('the pricing record a Usage read model stores', () => {
           costBasis: 'unpriced',
         }),
       /unpriced record carries a cost/,
-    );
-    assert.throws(
-      () =>
-        decodeModelCallPricingRecord({
-          ...projectModelCallPricingRecord(attempt({ costBasis: 'unpriced', costUsd: undefined })),
-          costBasis: 'priced',
-        }),
-      /priced record carries no cost/,
-    );
-    assert.throws(
-      () =>
-        decodeModelCallPricingRecord({
-          ...projectModelCallPricingRecord(attempt()),
-          usageBasis: 'missing',
-        }),
-      /missing usage but carries tokens/,
     );
   });
 });
