+++ packages/storage/src/__tests__/sqlite-usage-schema.test.ts
@@ -20,60 +20,13 @@
 import assert from 'node:assert/strict';
 import { DatabaseSync } from 'node:sqlite';
 import { test } from 'node:test';
-import {
-  MODEL_CALL_ATTEMPT_SCHEMA_VERSION,
-  type ModelCallAttempt,
-} from '@maka/core/model-call-attempt';
 import { migrateSqliteUsageDatabase } from '../sqlite-usage-schema.js';
-
-const NOW = 1_750_000_000_000;
-
-function wideAttempt(overrides: Partial<ModelCallAttempt> = {}): ModelCallAttempt {
-  return {
-    schemaVersion: MODEL_CALL_ATTEMPT_SCHEMA_VERSION,
-    logicalCallId: 'call-1',
-    attemptId: 'attempt-1',
-    traceId: 'trace-1',
-    sessionId: 'session-1',
-    runId: 'run-1',
-    turnId: 'turn-1',
-    step: 0,
-    attempt: 0,
-    callKind: 'main',
-    providerId: 'anthropic',
-    modelId: 'claude-opus-5',
-    startedAt: NOW - 1_000,
-    completedAt: NOW - 500,
-    latencyMs: 500,
-    status: 'completed',
-    usageBasis: 'reported',
-    inputTokens: 100,
-    outputTokens: 20,
-    costBasis: 'priced',
-    costUsd: 0.004,
-    promptComposition: { segments: [{ kind: 'messages', bytes: 4_096 }] },
-    // Sized like the real thing: the request observation is what made a stored
-    // row grow with the conversation rather than with spend.
-    requestObservation: {
-      schemaVersion: 1,
-      digest: `sha256:${'a'.repeat(64)}`,
-      bytes: 27_817,
-      segments: Array.from({ length: 64 }, (_, index) => ({
-        kind: 'tool_schema' as const,
-        index,
-        cacheable: true,
-        comparison: 'exact' as const,
-        digest: `sha256:${String(index).padStart(64, '0')}`,
-        bytes: 434,
-        label: `tool-${index}`,
-      })),
-    },
-    providerRequestId: 'req-1',
-    httpStatus: 200,
-    pricingRevision: 3,
-    ...overrides,
-  };
-}
+import {
+  MODEL_CALL_NOW as NOW,
+  MODEL_CALL_PRICING_ROW_KEYS,
+  storedModelCallRecord as storedRecord,
+  wideModelCallAttempt as wideAttempt,
+} from './fixtures/model-call-attempt.js';
 
 test('usage migration backfills Session identity for existing ledger rows', () => {
   const database = new DatabaseSync(':memory:');
@@ -144,23 +97,7 @@ test('usage migration narrows ledger rows to the fields a cost answer reads', ()
     migrateSqliteUsageDatabase(database);
 
     const narrowed = storedRecord(database, wide.attemptId);
-    assert.deepEqual(Object.keys(narrowed).sort(), [
-      'attemptId',
-      'callKind',
-      'completedAt',
-      'costBasis',
-      'costUsd',
-      'inputTokens',
-      'latencyMs',
-      'logicalCallId',
-      'modelId',
-      'outputTokens',
-      'providerId',
-      'sessionId',
-      'status',
-      'turnId',
-      'usageBasis',
-    ]);
+    assert.deepEqual(Object.keys(narrowed).sort(), MODEL_CALL_PRICING_ROW_KEYS);
     // Every number a Usage total is built from reads the same after the fold.
     assert.equal(narrowed.costUsd, 0.004);
     assert.equal(narrowed.inputTokens, 100);
@@ -229,13 +166,6 @@ test('usage migration narrows every row, not just the first page', () => {
     database.close();
   }
 });
-
-function storedRecord(database: DatabaseSync, attemptId: string): Record<string, unknown> {
-  const row = database
-    .prepare('SELECT record_json FROM usage_model_call_attempts WHERE attempt_id = ?')
-    .get(attemptId) as { record_json?: string } | undefined;
-  return JSON.parse(row?.record_json ?? '{}') as Record<string, unknown>;
-}
 
 function recordBytes(database: DatabaseSync, attemptId: string): number {
   return Number(
