+++ packages/storage/src/conversation-operational-state.ts
@@ -85,6 +85,10 @@ class SqliteConversationOperationalStateStore implements ConversationOperational
       database
         .prepare('DELETE FROM core_root_turn_start_rejections WHERE session_id = ?')
         .run(sessionId);
+      // Cascades this run's events and the Usage projection's checkpoints.
+      // `usage_model_call_attempts` is deliberately absent from this list:
+      // deleting a conversation must not erase its spend from all-time Usage
+      // totals, so those rows outlive the authority they were projected from.
       database.prepare('DELETE FROM core_agent_runs WHERE session_id = ?').run(sessionId);
       database
         .prepare('DELETE FROM core_client_capability_session_grants WHERE session_id = ?')
